test: split monolithic test_app into focused modules
Move coverage into domain-specific test files and adjust conftest;
restore full suite (217 tests) after modularization cleanup.
