+++ tests/test_persistence.py
@@ -1,0 +1,190 @@
+import json
+from pathlib import Path
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+from core.config_manager import (
+    SecretStore,
+    get_secret_with_fallback,
+    load_config,
+    save_config,
+    save_secret_with_fallback,
+)
+from core.models import AppConfig
+def test_secret_store_save_load():
+    with patch("core.config_manager.keyring") as mock_kr:
+        mock_kr.get_password.return_value = "STORED"
+        SecretStore.set_api_key("STORED")
+        mock_kr.set_password.assert_called_once_with(
+            "immich-go-gui", "default:api_key", "STORED"
+        )
+        assert SecretStore.get_api_key() == "STORED"
+
+def test_secret_store_migration():
+    with patch("core.config_manager.keyring") as mock_kr:
+        mock_kr.get_password.return_value = "OLD_KEY"
+        mock_settings = MagicMock()
+        mock_settings.value.return_value = "OLD_KEY"
+        SecretStore.migrate_from_qsettings(mock_settings)
+        mock_kr.set_password.assert_called_once_with(
+            "immich-go-gui", "default:api_key", "OLD_KEY"
+        )
+        mock_settings.remove.assert_called_once_with("api_key")
+
+def test_config_roundtrip(tmp_path, monkeypatch):
+    cfg_file = tmp_path / "config.toml"
+    monkeypatch.setenv("IMMICH_GO_GUI_CONFIG", str(cfg_file))
+
+    cfg = AppConfig()
+    cfg.server_url = "http://localhost:2283"
+    cfg.skip_ssl = True
+    cfg.allow_untested_updates = True
+
+    save_config(cfg)
+    loaded = load_config()
+
+    assert loaded.server_url == "http://localhost:2283"
+    assert loaded.skip_ssl is True
+    assert loaded.allow_untested_updates is True
+
+def test_secret_store_profile_scoped(monkeypatch):
+    store = {}
+
+    def mock_set(service, username, password):
+        store[username] = password
+
+    def mock_get(service, username):
+        return store.get(username, None)
+
+    def mock_delete(service, username):
+        store.pop(username, None)
+
+    monkeypatch.setattr("core.config_manager.keyring.set_password", mock_set)
+    monkeypatch.setattr("core.config_manager.keyring.get_password", mock_get)
+    monkeypatch.setattr("core.config_manager.keyring.delete_password", mock_delete)
+
+    assert SecretStore.set_secret("default", "api_key", "key_default") is True
+    assert SecretStore.set_secret("work", "api_key", "key_work") is True
+    assert SecretStore.set_secret("work", "admin_api_key", "admin_work") is True
+
+    assert SecretStore.get_secret("default", "api_key") == "key_default"
+    assert SecretStore.get_secret("work", "api_key") == "key_work"
+    assert SecretStore.get_secret("work", "admin_api_key") == "admin_work"
+
+    SecretStore.clear_secret("work", "api_key")
+    assert SecretStore.get_secret("work", "api_key") == ""
+    assert SecretStore.get_secret("default", "api_key") == "key_default"
+
+def test_secret_keyring_failure_fallback(tmp_path, monkeypatch):
+    secrets_file = tmp_path / "secrets.toml"
+
+    def mock_failing_set(service, username, password):
+        raise RuntimeError("Keyring unavailable")
+
+    def mock_failing_get(service, username):
+        return ""
+
+    monkeypatch.setattr("core.config_manager.keyring.set_password", mock_failing_set)
+    monkeypatch.setattr("core.config_manager.keyring.get_password", mock_failing_get)
+
+    res = save_secret_with_fallback(
+        profile_name="default",
+        key="api_key",
+        value="fallback_secret",
+        provider="keyring",
+        secrets_path=secrets_file,
+    )
+
+    assert res.ok is True
+    assert res.provider_used == "config"
+    assert "keyring is unavailable" in res.message.lower()
+
+    val = get_secret_with_fallback(
+        profile_name="default",
+        key="api_key",
+        provider="keyring",
+        secrets_path=secrets_file,
+    )
+    assert val == "fallback_secret"
+
+def test_collect_form_state_excludes_secrets(gui):
+    state = gui.collect_form_state()
+    for tab_name, tab_dict in state.items():
+        for secret_key in ("api_key", "from-api-key", "admin_api_key"):
+            assert secret_key not in tab_dict
+
+def test_secret_copy_success_verification(monkeypatch):
+    from core.config_manager import SecretStore
+
+    secrets_db = {}
+
+    def mock_set(profile, key, val):
+        secrets_db[(profile, key)] = val
+        return True
+
+    def mock_get(profile, key):
+        return secrets_db.get((profile, key), "")
+
+    monkeypatch.setattr(SecretStore, "set_secret", mock_set)
+    monkeypatch.setattr(SecretStore, "get_secret", mock_get)
+
+    secrets_db[("src", "api_key")] = "my-secret-key"
+    res = SecretStore.copy_secrets("src", "dst")
+    assert res is True
+    assert secrets_db.get(("dst", "api_key")) == "my-secret-key"
+
+def test_advanced_secret_value_not_persisted(gui):
+    gui.toggle_advanced(True)
+    gui.adv_rows["upload-immich"]["from-admin-api-key"].set_state(
+        {
+            "enabled": True,
+            "value": "super-secret-admin-key",
+        }
+    )
+
+    state = gui.collect_form_state()
+    saved = state["advanced"]["upload-immich"]["from-admin-api-key"]
+    assert saved["enabled"] is False
+    assert saved["value"] == ""
+
+def test_secret_status_label_keyring(gui, monkeypatch):
+    gui.app_config.secrets_provider = "keyring"
+    monkeypatch.setattr(
+        SecretStore, "get_secret", staticmethod(lambda *_: "secret-key")
+    )
+    monkeypatch.setattr(gui, "_secrets_file_has_key", lambda: False)
+    gui._update_secret_status()
+    assert "keyring" in gui.lbl_secret_status.text().lower()
+
+def test_secret_status_label_file_fallback(gui, monkeypatch):
+    gui.app_config.secrets_provider = "config"
+    monkeypatch.setattr(SecretStore, "get_secret", staticmethod(lambda *_: ""))
+    monkeypatch.setattr(gui, "_secrets_file_has_key", lambda: True)
+    gui._update_secret_status()
+    assert "secrets.toml" in gui.lbl_secret_status.text()
+
+def test_profile_index_cached(tmp_path, monkeypatch):
+    from core import profile_manager as pm
+
+    monkeypatch.setenv("IMMICH_GO_GUI_CONFIG", str(tmp_path / "config.toml"))
+    pm.clear_profiles_cache()
+    (tmp_path / "profiles.toml").write_text(
+        'schema_version = 1\nactive_profile = "default"\n[[profiles]]\nname = "default"\n',
+        encoding="utf-8",
+    )
+    # Point profiles path
+    monkeypatch.setattr(pm, "global_profiles_path", lambda: tmp_path / "profiles.toml")
+    pm.clear_profiles_cache()
+    first = pm._load_profiles_index()
+    (tmp_path / "profiles.toml").write_text(
+        'schema_version = 1\nactive_profile = "other"\n[[profiles]]\nname = "other"\n',
+        encoding="utf-8",
+    )
+    second = pm._load_profiles_index()
+    assert first is second
+    assert second.get("active_profile") == "default"
+    pm.clear_profiles_cache()
+    third = pm._load_profiles_index()
+    assert third.get("active_profile") == "other"
+
