+++ tests/test_widgets.py
@@ -1,0 +1,58 @@
+from unittest.mock import MagicMock, patch
+
+from PySide6.QtCore import QMimeData, QPointF, Qt, QUrl
+from PySide6.QtGui import QDropEvent
+from PySide6.QtWidgets import QFileDialog
+
+from gui.widgets import DroppablePlainTextEdit
+def test_droppable_plain_text_edit_drop(qapp, qtbot):
+    edit = DroppablePlainTextEdit()
+    qtbot.addWidget(edit)
+    mime = QMimeData()
+    mime.setUrls([QUrl.fromLocalFile("/path/a.zip"), QUrl.fromLocalFile("/path/b.zip")])
+    event = QDropEvent(
+        QPointF(0, 0),
+        Qt.DropAction.CopyAction,
+        mime,
+        Qt.MouseButton.LeftButton,
+        Qt.KeyboardModifier.NoModifier,
+    )
+    edit.dropEvent(event)
+    assert edit.toPlainText() == "/path/a.zip\n/path/b.zip"
+
+def test_browse_takeout_zips(gui):
+    gui.stacked_widget.setCurrentIndex(1)
+    gui.upload_tabs.setCurrentIndex(1)
+    with patch(
+        "PySide6.QtWidgets.QFileDialog.getOpenFileNames",
+        return_value=(["/path/a.zip", "/path/b.zip"], ""),
+    ):
+        gui.browse_takeout_zips()
+        assert (
+            gui.inputs["upload-gp"]["path"].toPlainText() == "/path/a.zip\n/path/b.zip"
+        )
+
+def test_browse_folder_upload(gui):
+    gui.stacked_widget.setCurrentIndex(1)
+    gui.upload_tabs.setCurrentIndex(0)
+    with patch(
+        "PySide6.QtWidgets.QFileDialog.getExistingDirectory",
+        return_value="/selected/folder",
+    ):
+        gui.browse_folder_upload()
+        assert gui.inputs["upload-folder"]["path"].text() == "/selected/folder"
+
+def test_native_dialog_options_passed(gui):
+    with patch(
+        "PySide6.QtWidgets.QFileDialog.getExistingDirectory", return_value="/test/path"
+    ) as mock_get_dir:
+        gui._browse_into(MagicMock(), "Test Caption")
+        mock_get_dir.assert_called_once()
+        from PySide6.QtWidgets import QFileDialog
+
+        args, kwargs = mock_get_dir.call_args
+        assert (
+            args[3] == QFileDialog.Option.ShowDirsOnly
+            or kwargs.get("options") == QFileDialog.Option.ShowDirsOnly
+        )
+
